feat: improve auth forms and landing page, update Supabase middleware
